sensor_temperature/requirements.txt
pandas
openpyxl==3.0.10
paho-mqtt==1.6.1
